requirements.txt
numpy
zmq
absl-py